backend/pyproject.toml
[project]
name = "memoria-os-backend"
version = "0.1.0"
description = "MemoriaOS Backend: The Multimodal Second Brain"
requires-python = ">=3.11"

dependencies = [
    # ── Google GenAI & ADK ───────────────────────────────────────────────
    "google-adk>=0.1.0",
    "google-genai>=0.1.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    
    # ── Storage & Backend ────────────────────────────────────────────────
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=2.18.2",
    "notion-client>=2.2.0",
    
    # ── Conversational & AI Tools ────────────────────────────────────────
    "google-cloud-texttospeech>=2.17.2",
    "httpx>=0.27.0",
    
    # ── Observability & Utils ────────────────────────────────────────────
    "structlog>=24.4.0",
    "python-dotenv>=1.0.0",
    "opentelemetry-api>=1.40.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "black",
    "ruff",
    "mypy",
]

[tool.uv]
dev-dependencies = [
    "memoria-os-backend[dev]",
    "pytest-mock>=3.15.1",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E","F","I","N","UP","ANN","S","B","SIM"]
ignore = ["ANN101","ANN102", "ANN201", "ANN401"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
